requirements.txt
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.4
cmake==3.29.3
contourpy==1.1.1
cycler==0.12.1
einops==0.7.0
filelock==3.13.1
fonttools==4.51.0
ftfy==6.2.0
gmpy2==2.1.2
idna==3.4
imageio==2.34.1
importlib_resources==6.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.4
lit==18.1.4
MarkupSafe==2.1.3
matplotlib==3.7.5
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.1
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
pip==23.3.1
protobuf==5.26.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
regex==2024.4.16
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
setuptools==68.2.2
six==1.16.0
sympy==1.12
tensorboardX==2.6.2.2
terminaltables==3.1.10
tifffile==2023.7.10
torch==2.0.0
torchvision==0.15.0
tqdm==4.66.2
triton==2.0.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
wheel==0.41.2
zipp==3.18.1
